Классы Order и OrderCancel разделены на класс работы с заказом и класс с проверками
